Refactor to have proper interface names - Hari
